refactor(api): extract inline handlers from createApp into route modules (cleanup 4/5)
createApp mixed ~150 loc of inline route handlers in with the middleware wiring.
Move them to three (ctx) => Hono modules on the existing mount loop:
- routes/oauth.ts    — OAuth/OIDC discovery (.well-known), consent + CLI-callback
                       screens, /v1/auth/providers
- routes/github-app.ts — the one-click GitHub App manifest flow (/new, /created)
- routes/system.ts   — /healthz, /readyz, the API-origin landing page

createApp drops 478 -> 319 lines and reads as pure composition: middleware,
host-dispatch (origin isolation + domain mode), CORS + rate limits, the auth
handler mount, the anonymous-write lockdown, then the router mount loop. All
moved handlers are GET / non-/v1 so the lockdown ordering is unaffected. No
behavior change; 618 tests green.
